refactor(LineSectionAcceptanceTest): fixture 클래스 분리
